zariz/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zariz-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi~=0.115",
  "uvicorn[standard]~=0.30",
  "pydantic~=2.9",
  "SQLAlchemy~=2.0",
  "psycopg2-binary~=2.9",
  "alembic~=1.13",
  "python-jose[cryptography]~=3.3",
  "passlib[bcrypt]~=1.7",
  "tenacity~=9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py312"]

